Replace PAGE_WHITE_LIST and SUFFIX_WHITE_LIST with WHITELIST_REGEXPS.
Using regexps we can also whitelist more complex URLs, e.g. for the
password reset tool (/passwordreset/1234...).

I chose to have a WHITELIST_REGEXPS_PATTERNS variable to hold the
patterns for the regular expressions since, in my opinion, such a list
is a bit easier to maintain.

WHITELIST_REGEXPS is a list of compiled regular expression
objects. Since we'll be using the regexps quite a lot, compiling them
should be a bit faster.

As for the patterns, I chose to have (most of) them preceded by a
slash and ending in a $. This way the effect is similar as the
previous implementation with the PAGE_WHITE_LIST variable. And it
seems to me that this is a bit more secure.