Bugfix. Get the minutes and hours for the date from the request.
Also, if the date is specified on the request, call hm_method with that date,
in case the minutes/hours are also specified in the date.

This is useful for when the form is called with a default date passed in via
the query string, such as when a VSEvent object is created in the Solgema.fullcalendar.