Fixed problem that occurs after upgrading the SearchableText index of the membrane_tool.
This happens after upgrading to membrane 2.0 or to Plone 4: the
membrane_tool catalog would be empty.  Now we refresh the
membrane_tool catalog when we upgrade the index.  If this has already
happened to you, it should work to just go to the membrane_tool, then
the Advanced tab, and click on 'Update Catalog.'